Added test file per server Change JSON match to JQ match Clean code
